# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OSM features API",
        "description": "The OSM Features API provides a comprehensive suite of endpoints for accessing OpenStreetMap (OSM) features in GeoJSON format, designed to streamline the integration and utilization of OSM data in applications. By leveraging daily syncs from the OSM database through osm2pgsql imports, the API ensures the delivery of up-to-date and efficiently processed data. Endpoints include retrieving lists and specific details of Places, Administrative Areas, Hiking Routes, and Poles, each identifiable by unique IDs and accompanied by their latest update timestamps. The API allows for fetching single items or lists, providing data in GeoJSON format that encompasses various attributes such as name, class, subclass, and geometry, making it versatile for numerous applications in mapping, navigation, and geographic analysis. This documentation aims to guide users through the API's functionalities, ensuring an understanding of how to make the most of the available resources for their mapping needs. Through concise and practical examples, users can learn how to query the API for specific geographic features and integrate these into their applications, enriching the user experience with detailed and dynamic map-based information.",
        "contact": {
            "email": "info@webmapp.it"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/features/places/list": {
            "get": {
                "tags": ["API V1"],
                "summary": "List all Places",
                "description": "Returns a list of Places with their IDs and updated timestamps. Optionally provide an 'updated_at' parameter to filter places updated after the specified date. Paginated results are available, with each page containing 100 places. Use the 'page' parameter to specify the page number to retrieve.",
                "operationId": "listPlaces",
                "parameters": [
                    {
                        "name": "updated_at",
                        "in": "query",
                        "description": "Places updated after this date",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number to retrieve. Each page contains 100 results.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PoiItem"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/features/places/{id}": {
            "get": {
                "tags": ["API V1"],
                "summary": "Get Place by ID",
                "description": "Returns a single Place in GeoJSON format",
                "operationId": "getPlaceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Place ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeoJsonFeature"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Place not found"
                    }
                }
            }
        },
        "/api/v1/features/admin-areas/list": {
            "get": {
                "tags": ["API V1"],
                "summary": "List all Admin Areas",
                "description": "Returns a list of Admin Areas with their details. Optionally, provide an 'updated_at' parameter to filter areas updated after the specified date. Paginated results are available, with each page containing 100 areas. Use the 'page' parameter to specify the page number to retrieve.",
                "operationId": "listAdminAreas",
                "parameters": [
                    {
                        "name": "updated_at",
                        "in": "query",
                        "description": "Admin Area updated after this date",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number to retrieve. Each page contains 100 results.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AdminAreaItem"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/features/admin-areas/{id}": {
            "get": {
                "tags": ["API V1"],
                "summary": "Get Admin Area by ID",
                "description": "Returns a single Admin Area in GeoJSON format",
                "operationId": "getAdminAreaById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Admin Area ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminAreaGeojsonFeature"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Admin Area not found"
                    }
                }
            }
        },
        "/api/v1/features/hiking-routes/list": {
            "get": {
                "tags": ["API V1"],
                "summary": "List all Hiking Routes",
                "description": "Returns a list of Hiking Routes with their details. Optionally, provide an 'updated_at' parameter to filter routes updated after the specified date. Paginated results are available, with each page containing 100 routes. Use the 'page' parameter to specify the page number to retrieve.",
                "operationId": "listHikingRoutes",
                "parameters": [
                    {
                        "name": "updated_at",
                        "in": "query",
                        "description": "Hiking Routes updated after this date",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number to retrieve. Each page contains 100 results.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HikingRouteItem"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/features/hiking-routes/{id}": {
            "get": {
                "tags": ["API V1"],
                "summary": "Get Hiking Routes by ID",
                "description": "Returns a single Hiking Route in GeoJSON format",
                "operationId": "getHikingRouteById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Hiking Route ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HikingRouteGeojsonFeature"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hiking Route not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "PoiItem": {
                "properties": {
                    "osm_id": {
                        "type": "integer",
                        "example": 123
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-03-10T02:00:00Z"
                    }
                },
                "type": "object"
            },
            "GeoJsonFeature": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "Feature"
                    },
                    "properties": {
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "Duomo di Arezzo"
                            },
                            "class": {
                                "type": "string",
                                "example": "amenity"
                            },
                            "subclass": {
                                "type": "string",
                                "example": "place_of_worship"
                            },
                            "osm_id": {
                                "type": "integer",
                                "example": 22720153
                            },
                            "osm_type": {
                                "type": "string",
                                "example": "W"
                            }
                        },
                        "type": "object"
                    },
                    "geometry": {
                        "properties": {
                            "type": {
                                "type": "string",
                                "example": "Point"
                            },
                            "coordinates": {
                                "type": "array",
                                "items": {
                                    "type": "number"
                                },
                                "example": [11.8805, 43.4635]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "AdminAreaItem": {
                "properties": {
                    "osm_id": {
                        "type": "integer",
                        "example": 39150
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-03-10T02:00:00Z"
                    }
                },
                "type": "object"
            },
            "AdminAreaGeojsonFeature": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "Feature"
                    },
                    "properties": {
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "Lampedusa e Linosa"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2021-03-10T02:00:00Z"
                            },
                            "admin_level": {
                                "type": "string",
                                "example": "8"
                            },
                            "osm_id": {
                                "type": "integer",
                                "example": 39150
                            },
                            "osm_type": {
                                "type": "string",
                                "example": "R"
                            },
                            "tags": {
                                "type": "json",
                                "example": "{\"name\": \"Lampedusa e Linosa\", \"type\": \"boundary\", \"name:ar\": \"لنبذوشة ونموشة\", \"name:es\": \"Lampedusa y Linosa\", \"name:it\": \"Lampedusa e Linosa\", \"name:la\": \"Lopadusa\", \"name:ru\": \"Лампедуза и Линоза\", \"boundary\": \"administrative\", \"wikidata\": \"Q224813\", \"ref:ISTAT\": \"084020\", \"wikipedia\": \"it:Lampedusa e Linosa\", \"admin_level\": \"8\", \"postal_code\": \"92031\", \"ref:catasto\": \"E431\"}"
                            }
                        },
                        "type": "object"
                    },
                    "geometry": {
                        "properties": {
                            "type": {
                                "type": "string",
                                "example": "Polygon"
                            },
                            "coordinates": {
                                "type": "array",
                                "items": {
                                    "type": "array",
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "type": "number"
                                        }
                                    }
                                },
                                "example": [
                                    [
                                        [11.8805, 43.4635],
                                        [11.8805, 43.4635],
                                        [11.8805, 43.4635],
                                        [11.8805, 43.4635]
                                    ]
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "HikingRouteItem": {
                "properties": {
                    "osm_id": {
                        "type": "integer",
                        "example": 39150
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-03-10T02:00:00Z"
                    }
                },
                "type": "object"
            },
            "HikingRouteGeojsonFeature": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "Feature"
                    },
                    "properties": {
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "Kyparissia Nature Trial"
                            },
                            "description": {
                                "type": "string",
                                "example": ""
                            },
                            "osm_id": {
                                "type": "integer",
                                "example": 39150
                            },
                            "osm_type": {
                                "type": "string",
                                "example": "R"
                            },
                            "tags": {
                                "type": "json",
                                "example": "{\"name\": \"Kyparissia Nature Trial\", \"type\": \"route\", \"route\": \"hiking\", \"hiking\": \"lwn\", \"name:en\": \"Kyparissia Nature Trial\", \"network\": \"lwn\", \"network:type\": \"lwn\"}"
                            }
                        },
                        "type": "object"
                    },
                    "geometry": {
                        "properties": {
                            "type": {
                                "type": "string",
                                "example": "MultilineString"
                            },
                            "coordinates": {
                                "type": "array",
                                "items": {
                                    "type": "array",
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "type": "number"
                                        }
                                    }
                                },
                                "example": [
                                    [
                                        [11.8805, 43.4635],
                                        [11.8805, 43.4635],
                                        [11.8805, 43.4635],
                                        [11.8805, 43.4635]
                                    ]
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        }
    }
}